Restricted tasks to roles.
I like restricting tasks to roles, partly out of laziness. I don't need
to specify HOSTS or ROLES on the command line as long as I've defined
the relevant roles somewhere. It also provides a safety net. Without
specifying HOSTS or ROLES on the command line, a task will get run on
*all* defined hosts. This can be avoided by restricting the tasks
themselves in deprec.

A problem with restricting tasks to roles in deprec was that Capistrano
complained when I ran a task in an ad-hoc manner without any hosts being
defined for the roles it was restricted to.

Sometimes I want to run a task on hosts but don't want to add them to
a deploy.rb. Editing a file just to use a command line is extra work!

e.g. cap deprec:mysql:install HOSTS=foo,bar

This problem caused me to remove a lot of role restrictions from deprec
over the last year or so.

My Capistrano patch (included in capistrano-2.5.0) causes Capistrano
to not complain about an undefined role if HOSTS or ROLES are specified
on the command line.

http://github.com/jamis/capistrano/commit/8badb0c522b38fae9cbe973451472afbddef61cb

So, I'm putting the roles back in.

- Mike